refactoring register account code
