+++ src/app/[locale]/(auth)/components/AuthCard.tsx
